+++ src/server/core/debug.cpp
@@ -183,9 +183,9 @@ void ShowServerStats(CONSOLE_CTX console)
 /**
  * Show queue stats from function
  */
-void ShowQueueStats(CONSOLE_CTX console, int size, const TCHAR *pszName)
+void ShowQueueStats(CONSOLE_CTX console, size_t size, const TCHAR *pszName)
 {
-   ConsolePrintf(console, _T("%-32s : %d\n"), pszName, size);
+   ConsolePrintf(console, _T("%-32s : %u\n"), pszName, static_cast<unsigned int>(size));
 }
 
 /**
